refactor(prism): moving node_types/sequencer to own crate
