+++ crates/prism/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod cfg;
 pub mod node_types;
 pub mod utils;
-pub mod webserver;
 #[macro_use]
 extern crate log;
