+++ crates/sequencer/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod sequencer;
+pub mod webserver;
+
+pub use sequencer::Sequencer;
+
+#[macro_use]
+extern crate log;
