+++ crates/sequencer/src/webserver.rs
@@ -1,4 +1,4 @@
-use crate::{cfg::WebServerConfig, node_types::sequencer::Sequencer};
+use crate::sequencer::Sequencer;
 use anyhow::{Context, Result};
 use axum::{
     extract::State,
@@ -22,6 +22,21 @@ use std::{self, sync::Arc};
 use tower_http::cors::CorsLayer;
 use utoipa::{OpenApi, ToSchema};
 use utoipa_swagger_ui::SwaggerUi;
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct WebServerConfig {
+    pub host: String,
+    pub port: u16,
+}
+
+impl Default for WebServerConfig {
+    fn default() -> Self {
+        WebServerConfig {
+            host: "127.0.0.1".to_string(),
+            port: 8089,
+        }
+    }
+}
 
 pub struct WebServer {
     pub cfg: WebServerConfig,
